refactor: replace bare unsigned with int throughout codebase
Replace all occurrences of bare `unsigned` (unsigned int) with `int`
across all headers and implementation files. Preserved unsigned char
(Color/Byte types), unsigned short/long (BMP binary format structs).

Added (int) casts where needed to fix -Wsign-compare warnings when
comparing int variables against container::size() return values.
Changed 1U/2U literals to 1/2 to resolve overload ambiguity after
removing unsigned overloads.

+++ code/game/controller/menu_controller.cc
@@ -25,7 +25,7 @@ std::string FullTextString() { return "full screen: "; }
 
 void MenuDisplay::Draw(ScalingDrawer * /*pDr*/) {
   Point p = pLeftTop;
-  for (unsigned i = 0; i < pCurr->vEntries.size(); ++i) {
+  for (int i = 0; i < (int)pCurr->vEntries.size(); ++i) {
     if (!pCurr->vEntries[i].bDisabled)
       pNum->DrawWord(pCurr->vEntries[i].sText, p, false);
     else
@@ -44,7 +44,7 @@ void MenuDisplay::OnMouseMove(Point pMouse) {
   pMouse.y /= 4;
 
   Point p = pLeftTop;
-  for (unsigned i = 0; i < pCurr->vEntries.size(); ++i) {
+  for (int i = 0; i < (int)pCurr->vEntries.size(); ++i) {
     if (!pCurr->vEntries[i].bDisabled &&
         InsideRectangle(Rectangle(p, pCurr->vEntries[i].szSize), pMouse)) {
       pCurr->nMenuPosition = i;
